+++ tests/test_user.py
@@ -1,5 +1,7 @@
+import pytest
 from laktory.models.resources.databricks import Group
 from laktory.models.resources.databricks import User
+from laktory.models.resources.databricks.user import UserLookup
 
 user = User(
     user_name="data.engineer@gmail.moc",
@@ -18,6 +20,15 @@ def test_user_group():
     assert user.roles == ["store-admin"]
     assert group.display_name == "role-engineers"
 
+def test_both_user_id_and_user_name():
+    with pytest.raises(ValueError):
+        UserLookup(user_id=123, user_name="test_user")    
+
+def test_neither_user_id_and_user_name():
+    with pytest.raises(ValueError):
+        UserLookup()
 
 if __name__ == "__main__":
     test_user_group()
+    test_both_user_id_and_user_name()
+    test_neither_user_id_and_user_name()